Move calculate_scores_for_validation functions to separate file
